Swap - Rework needed for AccountHeader actions + preselect account
Following the meeting for this functionality, this is the implementations of the agreed logic:

There will be no fetching of available providers until the user has accepted the KYC checkbox inside the Swap feature. This means that the CTA for swap will not be present in the accounts themselves until the user has accessed, accepted KYC and synced the providers at least once. This cache of providers is kept between sessions in the settings state instead of a volatile appstate to prevent polling the endpoint on every launch, and will be updated only when accessing the feature itself.

This work also includes the logic for preselecting the account when accessing the feature from an account screen as a source for a swap, as long as it's a valid account.