Allow iOS chained modals to show when accessing manager
onResult was being called as a side effect of setState, the modal wasn't hidden yet meaning that react is incapable of showing another modal. This was only happening on iOS due to faster native modals (or slower, not sure). Could be mitigated by an ugly setTimeout on the callback that shows the next modal, or by waiting for the onModalHide event, the later seems like the cleaner solution for now.