LL-8866 Move PanAndZoomView back to Animatied from react-native
Using reanimated was working on Android but not on iOS.
It would need further investigation to understand why it wasn't working
on iOS and how to fix it. For now this should be fine for this feature.
Once we move this view to the design system we can try once again reanimated