LIVE-1468 Prevent useless rerenders on NFT Gallery (#2257)
* LIVE-1468 Add RN performance monitor flipper plugin

* LIVE-1468 Implement rerender prevention quick-wins

This commit mainly adds memoing to a few components and values
and removes static function, objects and arrays declarations from the
render function

* LIVE-1468 Memoize NFT components as HOC to avoid rerenders

This commit implements a technique of splitting components into
presentational ones and context aware components. The presentational
components gets memoized so that they don't rerender at each context
change but rather only when the selected part of the context that
concerns them changes.

* LIVE-1468 Fix flow issues

* LIVE-1468 Fix flow issues

* LIVE-1468 Fix prettier issues